# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/photos/": {
            "get": {
                "tags": [
                    "photos"
                ],
                "summary": "List Photos",
                "operationId": "list_photos_photos__get",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "title": "Per Page"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response List Photos Photos  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "photos"
                ],
                "summary": "Upload Photo",
                "operationId": "upload_photo_photos__post",
                "parameters": [
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_photo_photos__post"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhotoPublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/photos/{photo_id}": {
            "get": {
                "tags": [
                    "photos"
                ],
                "summary": "Get Photo",
                "operationId": "get_photo_photos__photo_id__get",
                "parameters": [
                    {
                        "name": "photo_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Photo Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhotoPublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "photos"
                ],
                "summary": "Update Caption",
                "operationId": "update_caption_photos__photo_id__patch",
                "parameters": [
                    {
                        "name": "photo_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Photo Id"
                        }
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PhotoCaptionUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhotoPublic"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "photos"
                ],
                "summary": "Delete Photo",
                "operationId": "delete_photo_photos__photo_id__delete",
                "parameters": [
                    {
                        "name": "photo_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Photo Id"
                        }
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/photos/file/{filename}": {
            "get": {
                "tags": [
                    "photos"
                ],
                "summary": "Serve Photo",
                "operationId": "serve_photo_photos_file__filename__get",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Filename"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Body_upload_photo_photos__post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File"
                    },
                    "caption": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Caption"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_photo_photos__post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PhotoCaptionUpdate": {
                "properties": {
                    "caption": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 500
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Caption"
                    }
                },
                "type": "object",
                "title": "PhotoCaptionUpdate"
            },
            "PhotoPublic": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "original_filename": {
                        "type": "string",
                        "title": "Original Filename"
                    },
                    "caption": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Caption"
                    },
                    "uploaded_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Uploaded At"
                    },
                    "url": {
                        "type": "string",
                        "title": "Url"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "original_filename",
                    "caption",
                    "uploaded_at",
                    "url"
                ],
                "title": "PhotoPublic"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}